Extract /random page into own controller
Also fixes some rubocop-styleguide-violations.
