+++ config/routes.rb
@@ -7,7 +7,7 @@ Frontend::Application.routes.draw do
   post "/search" => proc { [405, {}, ["Method Not Allowed"]] } # Prevent non-GET requests for /search blowing up in the publication handlers below
   get "/search/opensearch" => "search#opensearch"
 
-  get "/random" => "root#random_page"
+  get "/random" => "random#random_page"
 
   # Crude way of handling the situation described at
   # http://stackoverflow.com/a/3443678
