+++ test/functional/random_controller_test.rb
@@ -1,0 +1,61 @@
+require 'test_helper'
+
+class RandomControllerTest < ActionController::TestCase
+  context "random page route" do
+    context 'valid search results' do
+      setup do
+        results = { "results" => [
+            { "link" => "/bereavement-allowance" },
+            { "link" => "/book-life-in-uk-test" },
+            { "link" => "http://www.wyreforestdc.gov.uk" },
+        ]}
+
+        stub_request(:get, "#{Plek.new.find('search')}/unified_search.json?count=0").to_return(status: 200, body: '{ "total": 3 }')
+        stub_request(:get, %r{#{Plek.new.find('search')}/unified_search.json\?count=1&fields=link&start=.}).to_return(status: 200, body: results.to_json)
+      end
+
+      should "redirect to one of the pages returned by search unless they start with http" do
+        get :random_page
+
+        expected_urls = [
+          "#{Plek.new.website_root}/bereavement-allowance",
+          "#{Plek.new.website_root}/book-life-in-uk-test",
+        ]
+        unexpected_urls = ["#{Plek.new.website_root}/http://www.wyreforestdc.gov.uk"]
+
+        assert_response :redirect
+        assert_includes(expected_urls, response.redirect_url)
+        assert_not_includes(unexpected_urls, response.redirect_url)
+      end
+
+      should "be cacheable long enough to discourage bots and short enough that users don't notice" do
+        get :random_page
+        assert_equal "max-age=5, public", response.headers["Cache-Control"]
+      end
+    end
+
+    context 'invalid search results' do
+      setup do
+        results = { "results" => [
+            { "link" => "http://www.wyreforestdc.gov.uk" },
+            { "link" => "/bereavement-allowance" },
+            { "link" => "/book-life-in-uk-test" },
+        ]}
+
+        stub_request(:get, "#{Plek.new.find('search')}/unified_search.json?count=0").to_return(status: 200, body: '{ "total": 3 }')
+        stub_request(:get, %r{#{Plek.new.find('search')}/unified_search.json\?count=1&fields=link&start=.}).to_return(status: 200, body: results.to_json)
+      end
+
+      should "not redirect to URLs that start with 'http' and start again" do
+        get :random_page
+
+        assert_response :redirect
+        assert_equal("#{Plek.new.website_root}/random", response.redirect_url)
+      end
+
+      should "not cache the bad case where it redirects to /random" do
+        assert_nil response.headers["Cache-Control"]
+      end
+    end
+  end
+end
